requirements.txt
Django==5.2.7
ultralytics==8.3.220
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
requests>=2.31.0
